+++ app_utils/caching.py
@@ -66,12 +66,9 @@ class ObjectCacheMixin:
             if select_related
             else ""
         )
-        return "{}_{}_{}{}".format(
-            self.model._meta.app_label,
-            self.model._meta.model_name,
-            pk,
-            f"_{suffix}" if suffix else "",
-        )
+        model_meta = self.model._meta
+        key = f"{model_meta.app_label}-{model_meta.model_name}-{pk}"
+        return f"{key}-{suffix}" if suffix else ""
 
     def _fetch_object_for_cache(self, pk, select_related: Optional[str] = None):
         qs = self.select_related(select_related) if select_related else self