+++ app_utils/json.py
@@ -24,10 +24,10 @@ class JSONDateTimeDecoder(json.JSONDecoder):
         :meta private:
         """
         json.JSONDecoder.__init__(
-            self, object_hook=self.dict_to_object, *args, **kwargs
+            self, object_hook=self._dict_to_object, *args, **kwargs
         )
 
-    def dict_to_object(self, dct: dict) -> object:
+    def _dict_to_object(self, dct: dict) -> object:
         if "__type__" not in dct:
             return dct
 
@@ -35,8 +35,8 @@ class JSONDateTimeDecoder(json.JSONDecoder):
         zone, _ = dct.pop("tz")
         dct["tzinfo"] = timezone(zone)
         try:
-            dateobj = dt.datetime(**dct)
-            return dateobj
+            date_obj = dt.datetime(**dct)
+            return date_obj
         except (ValueError, TypeError):
             dct["__type__"] = type_str
             return dct
@@ -70,5 +70,5 @@ class JSONDateTimeEncoder(json.JSONEncoder):
                 "microsecond": o.microsecond,
                 "tz": (o.tzinfo.tzname(o), o.utcoffset().total_seconds()),
             }
-        else:
-            return json.JSONEncoder.default(self, o)
+
+        return json.JSONEncoder.default(self, o)