+++ app_utils/logging.py
@@ -19,12 +19,15 @@ class LoggerAddTag(logging.LoggerAdapter):
 
     def __init__(self, my_logger, prefix):
         """:meta private:"""
-        super(LoggerAddTag, self).__init__(my_logger, {})
+        super().__init__(my_logger, {})
         self.prefix = prefix
 
     def process(self, msg, kwargs):
         """:meta private:"""
-        return "[%s] %s" % (self.prefix, msg), kwargs
+        return (
+            "[%s] %s" % (self.prefix, msg),  # pylint: disable=consider-using-f-string
+            kwargs,
+        )
 
 
 logger = LoggerAddTag(logging.getLogger(__name__), __package__)
@@ -32,4 +35,4 @@ logger = LoggerAddTag(logging.getLogger(__name__), __package__)
 
 def make_logger_prefix(tag: str):
     """creates a function to add logger prefixes, which returns tag when used empty"""
-    return lambda text="": "{}{}".format(tag, (": " + text) if text else "")
+    return lambda text="": f'{tag}{(": " + text) if text else ""}'